update room entity

+++ backend/src/main/java/com/mpbhms/backend/dto/AddRoomDTO.java
@@ -14,7 +14,7 @@ public class AddRoomDTO {
     private Long id;
     private String roomNumber;
     private BigDecimal area;
-    private BigDecimal pricePerMonth;
+    private Double pricePerMonth;
     private String roomStatus; // Enum: Available, Occupied, Maintenance, Inactive
     private Integer numberOfBedrooms;
     private Integer numberOfBathrooms;
